+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   When game updates happen, Hephaistos will now try to patch anyway (with warnings) instead of not patching (with errors).
+
 ## [1.6.0] - 2022-02-13
 
 ### Added